+++ packages/core/src/defer/instructions.ts
@@ -1,0 +1,757 @@
+/**
+ * @license
+ * Copyright Google LLC All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.io/license
+ */
+
+import {InjectionToken, Injector} from '../di';
+import {findMatchingDehydratedView} from '../hydration/views';
+import {populateDehydratedViewsInLContainer} from '../linker/view_container_ref';
+import {assertLContainer, assertTNodeForLView} from '../render3/assert';
+import {bindingUpdated} from '../render3/bindings';
+import {getComponentDef, getDirectiveDef, getPipeDef} from '../render3/definition';
+import {markViewDirty} from '../render3/instructions/mark_view_dirty';
+import {ɵɵtemplate} from '../render3/instructions/template';
+import {LContainer} from '../render3/interfaces/container';
+import {DirectiveDefList, PipeDefList} from '../render3/interfaces/definition';
+import {TContainerNode, TNode} from '../render3/interfaces/node';
+import {isDestroyed} from '../render3/interfaces/type_checks';
+import {HEADER_OFFSET, INJECTOR, LView, PARENT, TVIEW, TView} from '../render3/interfaces/view';
+import {getCurrentTNode, getLView, getSelectedTNode, getTView, nextBindingIndex} from '../render3/state';
+import {isPlatformBrowser} from '../render3/util/misc_utils';
+import {getConstant, getTNode} from '../render3/util/view_utils';
+import {addLViewToLContainer, createAndRenderEmbeddedLView, removeLViewFromLContainer, shouldAddViewToDom} from '../render3/view_manipulation';
+import {assertDefined, throwError} from '../util/assert';
+
+import {DeferBlockCleanupManager, invokeTDetailsCleanup, registerTDetailsCleanup} from './cleanup';
+import {onHover, onInteraction, onViewport, registerDomTrigger} from './dom_triggers';
+import {onIdle} from './idle_scheduler';
+import {DEFER_BLOCK_STATE, DeferBlockBehavior, DeferBlockConfig, DeferBlockDependencyInterceptor, DeferBlockInternalState, DeferBlockState, DeferBlockTriggers, DeferDependenciesLoadingState, DeferredLoadingBlockConfig, DeferredPlaceholderBlockConfig, DependencyResolverFn, LDeferBlockDetails, LOADING_AFTER_CLEANUP_FN, NEXT_DEFER_BLOCK_STATE, STATE_IS_FROZEN_UNTIL, TDeferBlockDetails} from './interfaces';
+import {onTimer, scheduleTimerTrigger} from './timer_scheduler';
+import {addDepsToRegistry, assertDeferredDependenciesLoaded, getLDeferBlockDetails, getLoadingBlockAfter, getMinimumDurationForState, getPrimaryBlockTNode, getTDeferBlockDetails, getTemplateIndexForState, setLDeferBlockDetails, setTDeferBlockDetails} from './utils';
+
+/**
+ * **INTERNAL**, avoid referencing it in application code.
+ *
+ * Injector token that allows to provide `DeferBlockDependencyInterceptor` class
+ * implementation.
+ */
+export const DEFER_BLOCK_DEPENDENCY_INTERCEPTOR =
+    new InjectionToken<DeferBlockDependencyInterceptor>(
+        ngDevMode ? 'DEFER_BLOCK_DEPENDENCY_INTERCEPTOR' : '');
+
+/**
+ * **INTERNAL**, token used for configuring defer block behavior.
+ */
+export const DEFER_BLOCK_CONFIG =
+    new InjectionToken<DeferBlockConfig>(ngDevMode ? 'DEFER_BLOCK_CONFIG' : '');
+
+/**
+ * Returns whether defer blocks should be triggered.
+ *
+ * Currently, defer blocks are not triggered on the server,
+ * only placeholder content is rendered (if provided).
+ */
+function shouldTriggerDeferBlock(injector: Injector): boolean {
+  const config = injector.get(DEFER_BLOCK_CONFIG, null, {optional: true});
+  if (config?.behavior === DeferBlockBehavior.Manual) {
+    return false;
+  }
+  return isPlatformBrowser(injector);
+}
+
+/**
+ * Reference to the timer-based scheduler implementation of defer block state
+ * rendering method. It's used to make timer-based scheduling tree-shakable.
+ * If `minimum` or `after` parameters are used, compiler generates an extra
+ * argument for the `ɵɵdefer` instruction, which references a timer-based
+ * implementation.
+ */
+let applyDeferBlockStateWithSchedulingImpl: (typeof applyDeferBlockState)|null = null;
+
+/**
+ * Enables timer-related scheduling if `after` or `minimum` parameters are setup
+ * on the `@loading` or `@placeholder` blocks.
+ */
+export function ɵɵdeferEnableTimerScheduling(
+    tView: TView, tDetails: TDeferBlockDetails, placeholderConfigIndex?: number|null,
+    loadingConfigIndex?: number|null) {
+  const tViewConsts = tView.consts;
+  if (placeholderConfigIndex != null) {
+    tDetails.placeholderBlockConfig =
+        getConstant<DeferredPlaceholderBlockConfig>(tViewConsts, placeholderConfigIndex);
+  }
+  if (loadingConfigIndex != null) {
+    tDetails.loadingBlockConfig =
+        getConstant<DeferredLoadingBlockConfig>(tViewConsts, loadingConfigIndex);
+  }
+
+  // Enable implementation that supports timer-based scheduling.
+  if (applyDeferBlockStateWithSchedulingImpl === null) {
+    applyDeferBlockStateWithSchedulingImpl = applyDeferBlockStateWithScheduling;
+  }
+}
+
+/**
+ * Creates runtime data structures for defer blocks.
+ *
+ * @param index Index of the `defer` instruction.
+ * @param primaryTmplIndex Index of the template with the primary block content.
+ * @param dependencyResolverFn Function that contains dependencies for this defer block.
+ * @param loadingTmplIndex Index of the template with the loading block content.
+ * @param placeholderTmplIndex Index of the template with the placeholder block content.
+ * @param errorTmplIndex Index of the template with the error block content.
+ * @param loadingConfigIndex Index in the constants array of the configuration of the loading.
+ *     block.
+ * @param placeholderConfigIndex Index in the constants array of the configuration of the
+ *     placeholder block.
+ * @param enableTimerScheduling Function that enables timer-related scheduling if `after`
+ *     or `minimum` parameters are setup on the `@loading` or `@placeholder` blocks.
+ *
+ * @codeGenApi
+ */
+export function ɵɵdefer(
+    index: number, primaryTmplIndex: number, dependencyResolverFn?: DependencyResolverFn|null,
+    loadingTmplIndex?: number|null, placeholderTmplIndex?: number|null,
+    errorTmplIndex?: number|null, loadingConfigIndex?: number|null,
+    placeholderConfigIndex?: number|null,
+    enableTimerScheduling?: typeof ɵɵdeferEnableTimerScheduling) {
+  const lView = getLView();
+  const tView = getTView();
+  const adjustedIndex = index + HEADER_OFFSET;
+
+  ɵɵtemplate(index, null, 0, 0);
+
+  if (tView.firstCreatePass) {
+    const tDetails: TDeferBlockDetails = {
+      primaryTmplIndex,
+      loadingTmplIndex: loadingTmplIndex ?? null,
+      placeholderTmplIndex: placeholderTmplIndex ?? null,
+      errorTmplIndex: errorTmplIndex ?? null,
+      placeholderBlockConfig: null,
+      loadingBlockConfig: null,
+      dependencyResolverFn: dependencyResolverFn ?? null,
+      loadingState: DeferDependenciesLoadingState.NOT_STARTED,
+      loadingPromise: null,
+    };
+    enableTimerScheduling?.(tView, tDetails, placeholderConfigIndex, loadingConfigIndex);
+    setTDeferBlockDetails(tView, adjustedIndex, tDetails);
+  }
+
+  const tNode = getCurrentTNode()!;
+  const lContainer = lView[adjustedIndex];
+
+  // If hydration is enabled, looks up dehydrated views in the DOM
+  // using hydration annotation info and stores those views on LContainer.
+  // In client-only mode, this function is a noop.
+  populateDehydratedViewsInLContainer(lContainer, tNode, lView);
+
+  // Init instance-specific defer details and store it.
+  const lDetails: LDeferBlockDetails = [
+    null,                             // NEXT_DEFER_BLOCK_STATE
+    DeferBlockInternalState.Initial,  // DEFER_BLOCK_STATE
+    null,                             // STATE_IS_FROZEN_UNTIL
+    null                              // LOADING_AFTER_CLEANUP_FN
+  ];
+  setLDeferBlockDetails(lView, adjustedIndex, lDetails);
+}
+
+/**
+ * Loads defer block dependencies when a trigger value becomes truthy.
+ * @codeGenApi
+ */
+export function ɵɵdeferWhen(rawValue: unknown) {
+  const lView = getLView();
+  const bindingIndex = nextBindingIndex();
+  if (bindingUpdated(lView, bindingIndex, rawValue)) {
+    const value = Boolean(rawValue);  // handle truthy or falsy values
+    const tNode = getSelectedTNode();
+    const lDetails = getLDeferBlockDetails(lView, tNode);
+    const renderedState = lDetails[DEFER_BLOCK_STATE];
+    if (value === false && renderedState === DeferBlockInternalState.Initial) {
+      // If nothing is rendered yet, render a placeholder (if defined).
+      renderPlaceholder(lView, tNode);
+    } else if (
+        value === true &&
+        (renderedState === DeferBlockInternalState.Initial ||
+         renderedState === DeferBlockState.Placeholder)) {
+      // The `when` condition has changed to `true`, trigger defer block loading
+      // if the block is either in initial (nothing is rendered) or a placeholder
+      // state.
+      triggerDeferBlock(lView, tNode);
+    }
+  }
+}
+
+/**
+ * Prefetches the deferred content when a value becomes truthy.
+ * @codeGenApi
+ */
+export function ɵɵdeferPrefetchWhen(rawValue: unknown) {
+  const lView = getLView();
+  const bindingIndex = nextBindingIndex();
+
+  if (bindingUpdated(lView, bindingIndex, rawValue)) {
+    const value = Boolean(rawValue);  // handle truthy or falsy values
+    const tView = lView[TVIEW];
+    const tNode = getSelectedTNode();
+    const tDetails = getTDeferBlockDetails(tView, tNode);
+    if (value === true && tDetails.loadingState === DeferDependenciesLoadingState.NOT_STARTED) {
+      // If loading has not been started yet, trigger it now.
+      triggerPrefetching(tDetails, lView);
+    }
+  }
+}
+
+/**
+ * Sets up logic to handle the `on idle` deferred trigger.
+ * @codeGenApi
+ */
+export function ɵɵdeferOnIdle() {
+  scheduleDelayedTrigger(onIdle);
+}
+
+/**
+ * Sets up logic to handle the `prefetch on idle` deferred trigger.
+ * @codeGenApi
+ */
+export function ɵɵdeferPrefetchOnIdle() {
+  scheduleDelayedPrefetching(onIdle, DeferBlockTriggers.OnIdle);
+}
+
+/**
+ * Sets up logic to handle the `on immediate` deferred trigger.
+ * @codeGenApi
+ */
+export function ɵɵdeferOnImmediate() {
+  const lView = getLView();
+  const tNode = getCurrentTNode()!;
+  const tView = lView[TVIEW];
+  const tDetails = getTDeferBlockDetails(tView, tNode);
+
+  // Render placeholder block only if loading template is not present
+  // to avoid content flickering, since it would be immediately replaced
+  // by the loading block.
+  if (tDetails.loadingTmplIndex === null) {
+    renderPlaceholder(lView, tNode);
+  }
+  triggerDeferBlock(lView, tNode);
+}
+
+
+/**
+ * Sets up logic to handle the `prefetch on immediate` deferred trigger.
+ * @codeGenApi
+ */
+export function ɵɵdeferPrefetchOnImmediate() {
+  const lView = getLView();
+  const tNode = getCurrentTNode()!;
+  const tView = lView[TVIEW];
+  const tDetails = getTDeferBlockDetails(tView, tNode);
+
+  if (tDetails.loadingState === DeferDependenciesLoadingState.NOT_STARTED) {
+    triggerResourceLoading(tDetails, lView);
+  }
+}
+
+/**
+ * Creates runtime data structures for the `on timer` deferred trigger.
+ * @param delay Amount of time to wait before loading the content.
+ * @codeGenApi
+ */
+export function ɵɵdeferOnTimer(delay: number) {
+  scheduleDelayedTrigger(onTimer(delay));
+}
+
+/**
+ * Creates runtime data structures for the `prefetch on timer` deferred trigger.
+ * @param delay Amount of time to wait before prefetching the content.
+ * @codeGenApi
+ */
+export function ɵɵdeferPrefetchOnTimer(delay: number) {
+  scheduleDelayedPrefetching(onTimer(delay), DeferBlockTriggers.OnTimer);
+}
+
+/**
+ * Creates runtime data structures for the `on hover` deferred trigger.
+ * @param triggerIndex Index at which to find the trigger element.
+ * @param walkUpTimes Number of times to walk up/down the tree hierarchy to find the trigger.
+ * @codeGenApi
+ */
+export function ɵɵdeferOnHover(triggerIndex: number, walkUpTimes?: number) {
+  const lView = getLView();
+  const tNode = getCurrentTNode()!;
+
+  renderPlaceholder(lView, tNode);
+  registerDomTrigger(
+      lView, tNode, triggerIndex, walkUpTimes, onHover, () => triggerDeferBlock(lView, tNode));
+}
+
+/**
+ * Creates runtime data structures for the `prefetch on hover` deferred trigger.
+ * @param triggerIndex Index at which to find the trigger element.
+ * @param walkUpTimes Number of times to walk up/down the tree hierarchy to find the trigger.
+ * @codeGenApi
+ */
+export function ɵɵdeferPrefetchOnHover(triggerIndex: number, walkUpTimes?: number) {
+  const lView = getLView();
+  const tNode = getCurrentTNode()!;
+  const tView = lView[TVIEW];
+  const tDetails = getTDeferBlockDetails(tView, tNode);
+
+  if (tDetails.loadingState === DeferDependenciesLoadingState.NOT_STARTED) {
+    registerDomTrigger(
+        lView, tNode, triggerIndex, walkUpTimes, onHover,
+        () => triggerPrefetching(tDetails, lView));
+  }
+}
+
+/**
+ * Creates runtime data structures for the `on interaction` deferred trigger.
+ * @param triggerIndex Index at which to find the trigger element.
+ * @param walkUpTimes Number of times to walk up/down the tree hierarchy to find the trigger.
+ * @codeGenApi
+ */
+export function ɵɵdeferOnInteraction(triggerIndex: number, walkUpTimes?: number) {
+  const lView = getLView();
+  const tNode = getCurrentTNode()!;
+
+  renderPlaceholder(lView, tNode);
+  registerDomTrigger(
+      lView, tNode, triggerIndex, walkUpTimes, onInteraction,
+      () => triggerDeferBlock(lView, tNode));
+}
+
+/**
+ * Creates runtime data structures for the `prefetch on interaction` deferred trigger.
+ * @param triggerIndex Index at which to find the trigger element.
+ * @param walkUpTimes Number of times to walk up/down the tree hierarchy to find the trigger.
+ * @codeGenApi
+ */
+export function ɵɵdeferPrefetchOnInteraction(triggerIndex: number, walkUpTimes?: number) {
+  const lView = getLView();
+  const tNode = getCurrentTNode()!;
+  const tView = lView[TVIEW];
+  const tDetails = getTDeferBlockDetails(tView, tNode);
+
+  if (tDetails.loadingState === DeferDependenciesLoadingState.NOT_STARTED) {
+    registerDomTrigger(
+        lView, tNode, triggerIndex, walkUpTimes, onInteraction,
+        () => triggerPrefetching(tDetails, lView));
+  }
+}
+
+/**
+ * Creates runtime data structures for the `on viewport` deferred trigger.
+ * @param triggerIndex Index at which to find the trigger element.
+ * @param walkUpTimes Number of times to walk up/down the tree hierarchy to find the trigger.
+ * @codeGenApi
+ */
+export function ɵɵdeferOnViewport(triggerIndex: number, walkUpTimes?: number) {
+  const lView = getLView();
+  const tNode = getCurrentTNode()!;
+
+  renderPlaceholder(lView, tNode);
+  registerDomTrigger(
+      lView, tNode, triggerIndex, walkUpTimes, onViewport, () => triggerDeferBlock(lView, tNode));
+}
+
+/**
+ * Creates runtime data structures for the `prefetch on viewport` deferred trigger.
+ * @param triggerIndex Index at which to find the trigger element.
+ * @param walkUpTimes Number of times to walk up/down the tree hierarchy to find the trigger.
+ * @codeGenApi
+ */
+export function ɵɵdeferPrefetchOnViewport(triggerIndex: number, walkUpTimes?: number) {
+  const lView = getLView();
+  const tNode = getCurrentTNode()!;
+  const tView = lView[TVIEW];
+  const tDetails = getTDeferBlockDetails(tView, tNode);
+
+  if (tDetails.loadingState === DeferDependenciesLoadingState.NOT_STARTED) {
+    registerDomTrigger(
+        lView, tNode, triggerIndex, walkUpTimes, onViewport,
+        () => triggerPrefetching(tDetails, lView));
+  }
+}
+
+/********** Helper functions **********/
+
+/**
+ * Schedules triggering of a defer block for `on idle` and `on timer` conditions.
+ */
+function scheduleDelayedTrigger(
+    scheduleFn: (callback: VoidFunction, lView: LView, withLViewCleanup: boolean) => VoidFunction) {
+  const lView = getLView();
+  const tNode = getCurrentTNode()!;
+
+  renderPlaceholder(lView, tNode);
+  scheduleFn(() => triggerDeferBlock(lView, tNode), lView, true /* withLViewCleanup */);
+}
+
+/**
+ * Schedules prefetching for `on idle` and `on timer` triggers.
+ *
+ * @param scheduleFn A function that does the scheduling.
+ * @param trigger A trigger that initiated scheduling.
+ */
+function scheduleDelayedPrefetching(
+    scheduleFn: (callback: VoidFunction, lView: LView, withLViewCleanup: boolean) => VoidFunction,
+    trigger: DeferBlockTriggers) {
+  const lView = getLView();
+  const tNode = getCurrentTNode()!;
+  const tView = lView[TVIEW];
+  const tDetails = getTDeferBlockDetails(tView, tNode);
+
+  if (tDetails.loadingState === DeferDependenciesLoadingState.NOT_STARTED) {
+    // Prevent scheduling more than one prefetch init call
+    // for each defer block. For this reason we use only a trigger
+    // identifier in a key, so all instances would use the same key.
+    const key = String(trigger);
+    const injector = lView[INJECTOR]!;
+    const manager = injector.get(DeferBlockCleanupManager);
+    if (!manager.has(tDetails, key)) {
+      // In case of prefetching, we intentionally avoid cancelling resource loading if
+      // an underlying LView get destroyed (thus passing `null` as a second argument),
+      // because there might be other LViews (that represent embedded views) that
+      // depend on resource loading.
+      const prefetch = () => triggerPrefetching(tDetails, lView);
+      const cleanupFn = scheduleFn(prefetch, lView, false /* withLViewCleanup */);
+      registerTDetailsCleanup(injector, tDetails, key, cleanupFn);
+    }
+  }
+}
+
+/**
+ * Transitions a defer block to the new state. Updates the  necessary
+ * data structures and renders corresponding block.
+ *
+ * @param newState New state that should be applied to the defer block.
+ * @param tNode TNode that represents a defer block.
+ * @param lContainer Represents an instance of a defer block.
+ */
+export function renderDeferBlockState(
+    newState: DeferBlockState, tNode: TNode, lContainer: LContainer): void {
+  const hostLView = lContainer[PARENT];
+  const hostTView = hostLView[TVIEW];
+
+  // Check if this view is not destroyed. Since the loading process was async,
+  // the view might end up being destroyed by the time rendering happens.
+  if (isDestroyed(hostLView)) return;
+
+  // Make sure this TNode belongs to TView that represents host LView.
+  ngDevMode && assertTNodeForLView(tNode, hostLView);
+
+  const lDetails = getLDeferBlockDetails(hostLView, tNode);
+
+  ngDevMode && assertDefined(lDetails, 'Expected a defer block state defined');
+
+  const currentState = lDetails[DEFER_BLOCK_STATE];
+
+  if (isValidStateChange(currentState, newState) &&
+      isValidStateChange(lDetails[NEXT_DEFER_BLOCK_STATE] ?? -1, newState)) {
+    const tDetails = getTDeferBlockDetails(hostTView, tNode);
+    const needsScheduling = getLoadingBlockAfter(tDetails) !== null ||
+        getMinimumDurationForState(tDetails, DeferBlockState.Loading) !== null ||
+        getMinimumDurationForState(tDetails, DeferBlockState.Placeholder);
+
+    if (ngDevMode && needsScheduling) {
+      assertDefined(
+          applyDeferBlockStateWithSchedulingImpl, 'Expected scheduling function to be defined');
+    }
+
+    const applyStateFn =
+        needsScheduling ? applyDeferBlockStateWithSchedulingImpl! : applyDeferBlockState;
+    applyStateFn(newState, lDetails, lContainer, tNode, hostLView);
+  }
+}
+
+/**
+ * Applies changes to the DOM to reflect a given state.
+ */
+function applyDeferBlockState(
+    newState: DeferBlockState, lDetails: LDeferBlockDetails, lContainer: LContainer, tNode: TNode,
+    hostLView: LView<unknown>) {
+  const stateTmplIndex = getTemplateIndexForState(newState, hostLView, tNode);
+
+  if (stateTmplIndex !== null) {
+    lDetails[DEFER_BLOCK_STATE] = newState;
+    const hostTView = hostLView[TVIEW];
+    const adjustedIndex = stateTmplIndex + HEADER_OFFSET;
+    const tNode = getTNode(hostTView, adjustedIndex) as TContainerNode;
+
+    // There is only 1 view that can be present in an LContainer that
+    // represents a defer block, so always refer to the first one.
+    const viewIndex = 0;
+
+    removeLViewFromLContainer(lContainer, viewIndex);
+    const dehydratedView = findMatchingDehydratedView(lContainer, tNode.tView!.ssrId);
+    const embeddedLView = createAndRenderEmbeddedLView(hostLView, tNode, null, {dehydratedView});
+    addLViewToLContainer(
+        lContainer, embeddedLView, viewIndex, shouldAddViewToDom(tNode, dehydratedView));
+    markViewDirty(embeddedLView);
+  }
+}
+
+/**
+ * Extends the `applyDeferBlockState` with timer-based scheduling.
+ * This function becomes available on a page if there are defer blocks
+ * that use `after` or `minimum` parameters in the `@loading` or
+ * `@placeholder` blocks.
+ */
+function applyDeferBlockStateWithScheduling(
+    newState: DeferBlockState, lDetails: LDeferBlockDetails, lContainer: LContainer, tNode: TNode,
+    hostLView: LView<unknown>) {
+  const now = Date.now();
+  const hostTView = hostLView[TVIEW];
+  const tDetails = getTDeferBlockDetails(hostTView, tNode);
+
+  if (lDetails[STATE_IS_FROZEN_UNTIL] === null || lDetails[STATE_IS_FROZEN_UNTIL] <= now) {
+    lDetails[STATE_IS_FROZEN_UNTIL] = null;
+
+    const loadingAfter = getLoadingBlockAfter(tDetails);
+    const inLoadingAfterPhase = lDetails[LOADING_AFTER_CLEANUP_FN] !== null;
+    if (newState === DeferBlockState.Loading && loadingAfter !== null && !inLoadingAfterPhase) {
+      // Trying to render loading, but it has an `after` config,
+      // so schedule an update action after a timeout.
+      lDetails[NEXT_DEFER_BLOCK_STATE] = newState;
+      const cleanupFn =
+          scheduleDeferBlockUpdate(loadingAfter, lDetails, tNode, lContainer, hostLView);
+      lDetails[LOADING_AFTER_CLEANUP_FN] = cleanupFn;
+    } else {
+      // If we transition to a complete or an error state and there is a pending
+      // operation to render loading after a timeout - invoke a cleanup operation,
+      // which stops the timer.
+      if (newState > DeferBlockState.Loading && inLoadingAfterPhase) {
+        lDetails[LOADING_AFTER_CLEANUP_FN]!();
+        lDetails[LOADING_AFTER_CLEANUP_FN] = null;
+        lDetails[NEXT_DEFER_BLOCK_STATE] = null;
+      }
+
+      applyDeferBlockState(newState, lDetails, lContainer, tNode, hostLView);
+
+      const duration = getMinimumDurationForState(tDetails, newState);
+      if (duration !== null) {
+        lDetails[STATE_IS_FROZEN_UNTIL] = now + duration;
+        scheduleDeferBlockUpdate(duration, lDetails, tNode, lContainer, hostLView);
+      }
+    }
+  } else {
+    // We are still rendering the previous state.
+    // Update the `NEXT_DEFER_BLOCK_STATE`, which would be
+    // picked up once it's time to transition to the next state.
+    lDetails[NEXT_DEFER_BLOCK_STATE] = newState;
+  }
+}
+
+/**
+ * Schedules an update operation after a specified timeout.
+ */
+function scheduleDeferBlockUpdate(
+    timeout: number, lDetails: LDeferBlockDetails, tNode: TNode, lContainer: LContainer,
+    hostLView: LView<unknown>): VoidFunction {
+  const callback = () => {
+    const nextState = lDetails[NEXT_DEFER_BLOCK_STATE];
+    lDetails[STATE_IS_FROZEN_UNTIL] = null;
+    lDetails[NEXT_DEFER_BLOCK_STATE] = null;
+    if (nextState !== null) {
+      renderDeferBlockState(nextState, tNode, lContainer);
+    }
+  };
+  return scheduleTimerTrigger(timeout, callback, hostLView, true);
+}
+
+/**
+ * Checks whether we can transition to the next state.
+ *
+ * We transition to the next state if the previous state was represented
+ * with a number that is less than the next state. For example, if the current
+ * state is "loading" (represented as `1`), we should not show a placeholder
+ * (represented as `0`), but we can show a completed state (represented as `2`)
+ * or an error state (represented as `3`).
+ */
+function isValidStateChange(
+    currentState: DeferBlockState|DeferBlockInternalState, newState: DeferBlockState): boolean {
+  return currentState < newState;
+}
+
+/**
+ * Trigger prefetching of dependencies for a defer block.
+ *
+ * @param tDetails Static information about this defer block.
+ * @param lView LView of a host view.
+ */
+export function triggerPrefetching(tDetails: TDeferBlockDetails, lView: LView) {
+  if (lView[INJECTOR] && shouldTriggerDeferBlock(lView[INJECTOR]!)) {
+    triggerResourceLoading(tDetails, lView);
+  }
+}
+
+/**
+ * Trigger loading of defer block dependencies if the process hasn't started yet.
+ *
+ * @param tDetails Static information about this defer block.
+ * @param lView LView of a host view.
+ */
+export function triggerResourceLoading(tDetails: TDeferBlockDetails, lView: LView) {
+  const injector = lView[INJECTOR]!;
+  const tView = lView[TVIEW];
+
+  if (tDetails.loadingState !== DeferDependenciesLoadingState.NOT_STARTED) {
+    // If the loading status is different from initial one, it means that
+    // the loading of dependencies is in progress and there is nothing to do
+    // in this function. All details can be obtained from the `tDetails` object.
+    return;
+  }
+
+  const primaryBlockTNode = getPrimaryBlockTNode(tView, tDetails);
+
+  // Switch from NOT_STARTED -> IN_PROGRESS state.
+  tDetails.loadingState = DeferDependenciesLoadingState.IN_PROGRESS;
+
+  // Check if dependency function interceptor is configured.
+  const deferDependencyInterceptor =
+      injector.get(DEFER_BLOCK_DEPENDENCY_INTERCEPTOR, null, {optional: true});
+
+  const dependenciesFn = deferDependencyInterceptor ?
+      deferDependencyInterceptor.intercept(tDetails.dependencyResolverFn) :
+      tDetails.dependencyResolverFn;
+
+  // The `dependenciesFn` might be `null` when all dependencies within
+  // a given defer block were eagerly references elsewhere in a file,
+  // thus no dynamic `import()`s were produced.
+  if (!dependenciesFn) {
+    tDetails.loadingPromise = Promise.resolve().then(() => {
+      tDetails.loadingState = DeferDependenciesLoadingState.COMPLETE;
+    });
+    return;
+  }
+
+  // Defer block may have multiple prefetch triggers. Once the loading
+  // starts, invoke all clean functions, since they are no longer needed.
+  invokeTDetailsCleanup(injector, tDetails);
+
+  // Start downloading of defer block dependencies.
+  tDetails.loadingPromise = Promise.allSettled(dependenciesFn()).then(results => {
+    let failed = false;
+    const directiveDefs: DirectiveDefList = [];
+    const pipeDefs: PipeDefList = [];
+
+    for (const result of results) {
+      if (result.status === 'fulfilled') {
+        const dependency = result.value;
+        const directiveDef = getComponentDef(dependency) || getDirectiveDef(dependency);
+        if (directiveDef) {
+          directiveDefs.push(directiveDef);
+        } else {
+          const pipeDef = getPipeDef(dependency);
+          if (pipeDef) {
+            pipeDefs.push(pipeDef);
+          }
+        }
+      } else {
+        failed = true;
+        break;
+      }
+    }
+
+    // Loading is completed, we no longer need this Promise.
+    tDetails.loadingPromise = null;
+
+    if (failed) {
+      tDetails.loadingState = DeferDependenciesLoadingState.FAILED;
+    } else {
+      tDetails.loadingState = DeferDependenciesLoadingState.COMPLETE;
+
+      // Update directive and pipe registries to add newly downloaded dependencies.
+      const primaryBlockTView = primaryBlockTNode.tView!;
+      if (directiveDefs.length > 0) {
+        primaryBlockTView.directiveRegistry =
+            addDepsToRegistry<DirectiveDefList>(primaryBlockTView.directiveRegistry, directiveDefs);
+      }
+      if (pipeDefs.length > 0) {
+        primaryBlockTView.pipeRegistry =
+            addDepsToRegistry<PipeDefList>(primaryBlockTView.pipeRegistry, pipeDefs);
+      }
+    }
+  });
+}
+
+/** Utility function to render placeholder content (if present) */
+function renderPlaceholder(lView: LView, tNode: TNode) {
+  const lContainer = lView[tNode.index];
+  ngDevMode && assertLContainer(lContainer);
+
+  renderDeferBlockState(DeferBlockState.Placeholder, tNode, lContainer);
+}
+
+/**
+ * Subscribes to the "loading" Promise and renders corresponding defer sub-block,
+ * based on the loading results.
+ *
+ * @param lContainer Represents an instance of a defer block.
+ * @param tNode Represents defer block info shared across all instances.
+ */
+function renderDeferStateAfterResourceLoading(
+    tDetails: TDeferBlockDetails, tNode: TNode, lContainer: LContainer) {
+  ngDevMode &&
+      assertDefined(
+          tDetails.loadingPromise, 'Expected loading Promise to exist on this defer block');
+
+  tDetails.loadingPromise!.then(() => {
+    if (tDetails.loadingState === DeferDependenciesLoadingState.COMPLETE) {
+      ngDevMode && assertDeferredDependenciesLoaded(tDetails);
+
+      // Everything is loaded, show the primary block content
+      renderDeferBlockState(DeferBlockState.Complete, tNode, lContainer);
+
+    } else if (tDetails.loadingState === DeferDependenciesLoadingState.FAILED) {
+      renderDeferBlockState(DeferBlockState.Error, tNode, lContainer);
+    }
+  });
+}
+
+/**
+ * Attempts to trigger loading of defer block dependencies.
+ * If the block is already in a loading, completed or an error state -
+ * no additional actions are taken.
+ */
+function triggerDeferBlock(lView: LView, tNode: TNode) {
+  const tView = lView[TVIEW];
+  const lContainer = lView[tNode.index];
+  const injector = lView[INJECTOR]!;
+  ngDevMode && assertLContainer(lContainer);
+
+  if (!shouldTriggerDeferBlock(injector)) return;
+
+  const tDetails = getTDeferBlockDetails(tView, tNode);
+  switch (tDetails.loadingState) {
+    case DeferDependenciesLoadingState.NOT_STARTED:
+      renderDeferBlockState(DeferBlockState.Loading, tNode, lContainer);
+      triggerResourceLoading(tDetails, lView);
+
+      // The `loadingState` might have changed to "loading".
+      if ((tDetails.loadingState as DeferDependenciesLoadingState) ===
+          DeferDependenciesLoadingState.IN_PROGRESS) {
+        renderDeferStateAfterResourceLoading(tDetails, tNode, lContainer);
+      }
+      break;
+    case DeferDependenciesLoadingState.IN_PROGRESS:
+      renderDeferBlockState(DeferBlockState.Loading, tNode, lContainer);
+      renderDeferStateAfterResourceLoading(tDetails, tNode, lContainer);
+      break;
+    case DeferDependenciesLoadingState.COMPLETE:
+      ngDevMode && assertDeferredDependenciesLoaded(tDetails);
+      renderDeferBlockState(DeferBlockState.Complete, tNode, lContainer);
+      break;
+    case DeferDependenciesLoadingState.FAILED:
+      renderDeferBlockState(DeferBlockState.Error, tNode, lContainer);
+      break;
+    default:
+      if (ngDevMode) {
+        throwError('Unknown defer block state');
+      }
+  }
+}
