+++ packages/core/src/defer/interfaces.ts
@@ -6,7 +6,7 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import type {DependencyType} from './definition';
+import type {DependencyType} from '../render3/interfaces/definition';
 
 /**
  * Describes the shape of a function generated by the compiler
@@ -203,4 +203,24 @@ export enum DeferBlockBehavior {
    * Playthrough mode for defer blocks. This mode behaves like defer blocks would in a browser.
    */
   Playthrough,
+}
+
+/**
+ * **INTERNAL**, avoid referencing it in application code.
+ *
+ * Describes a helper class that allows to intercept a call to retrieve current
+ * dependency loading function and replace it with a different implementation.
+ * This interceptor class is needed to allow testing blocks in different states
+ * by simulating loading response.
+ */
+export interface DeferBlockDependencyInterceptor {
+  /**
+   * Invoked for each defer block when dependency loading function is accessed.
+   */
+  intercept(dependencyFn: DependencyResolverFn|null): DependencyResolverFn|null;
+
+  /**
+   * Allows to configure an interceptor function.
+   */
+  setInterceptor(interceptorFn: (current: DependencyResolverFn) => DependencyResolverFn): void;
 }
