+++ packages/core/src/render3/index.ts
@@ -160,11 +160,14 @@ export {
   ɵsetUnknownElementStrictMode,
   ɵgetUnknownPropertyStrictMode,
   ɵsetUnknownPropertyStrictMode,
-
-  DeferBlockDependencyInterceptor as ɵDeferBlockDependencyInterceptor,
+} from './instructions/all';
+export {
   DEFER_BLOCK_DEPENDENCY_INTERCEPTOR as ɵDEFER_BLOCK_DEPENDENCY_INTERCEPTOR,
   DEFER_BLOCK_CONFIG as ɵDEFER_BLOCK_CONFIG,
-} from './instructions/all';
+} from '../defer/instructions';
+export {
+  DeferBlockDependencyInterceptor as ɵDeferBlockDependencyInterceptor,
+} from '../defer/interfaces';
 export {ɵɵi18n, ɵɵi18nApply, ɵɵi18nAttributes, ɵɵi18nEnd, ɵɵi18nExp,ɵɵi18nPostprocess, ɵɵi18nStart} from './instructions/i18n';
 export {RenderFlags} from './interfaces/definition';
 export {
